fix: dispose provider kernels and share transport modes
