+++ packages/svelte/src/lib/index.ts
@@ -34,6 +34,10 @@ export type {
 	PolicyDecision,
 	PolicyScopeMode,
 	PolicyUiSurfaceConfig,
+	HostedModeOptions,
+	ProviderTransportContext,
+	ProviderTransportFactory,
+	ProviderTransportKind,
 	ResolvedPolicy,
 	SavePayload,
 	SaveResult,
@@ -91,13 +95,6 @@ export type {
 	UsePersistenceOptions,
 	UseScriptLoaderOptions,
 } from './types';
-export { custom } from './transports/custom';
-export type { HostedModeOptions } from './transports/hosted';
-export { hosted } from './transports/hosted';
+export { custom, hosted } from '@c15t/core/v3';
 export type { OfflineModeOptions } from './transports/offline';
 export { offline } from './transports/offline';
-export type {
-	ProviderTransportContext,
-	ProviderTransportFactory,
-	ProviderTransportKind,
-} from './transports/types';
