+++ packages/svelte/src/lib/server/index.ts
@@ -106,6 +106,5 @@ export const prefetchInitialConsent = async function prefetchInitialConsent(
 
 export type { KernelConfig } from '@c15t/core/v3';
 export type { PrefetchInitialConsentOptions, ReadInitialConsentConfigOptions };
-export { custom } from '../transports/custom';
-export { hosted } from '../transports/hosted';
+export { custom, hosted } from '@c15t/core/v3';
 export { offline } from '../transports/offline';
